working on navigationbuttons/layout utils by variant
