+++ app/src/main/res/layout/dynamic_tab_grid_question.xml
@@ -2,32 +2,23 @@
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    xmlns:layout="http://schemas.android.com/tools"
     android:background="@color/questionLayoutBackground">
 
-    <!--Horizontal Progress bar + text-->
-    <RelativeLayout
-        android:id="@+id/dynamic_footer"
-        android:layout_width="match_parent"
-        android:layout_height="30dp"
-        android:layout_alignParentBottom="true">
-        <include layout="@layout/progress_bar_row"></include>
-
-    </RelativeLayout>
 
     <LinearLayout
-        android:layout_above="@id/dynamic_footer"
+        android:id="@+id/dynamic_tab"
+        android:layout_alignParentTop="true"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
+        android:layout_above="@+id/multi_question_tab_navigation_buttons"
         android:orientation="vertical">
 
         <!--Question-->
         <include layout="@layout/question_header"></include>
-
-        <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical">
+            <LinearLayout
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:orientation="vertical">
             <ImageView
                 android:id="@+id/questionImage"
                 android:layout_width="match_parent"
@@ -46,7 +37,7 @@
                 <include layout="@layout/no_scrolled_table"></include>
 
 
-                <TableLayout xmlns:android="http://schemas.android.com/apk/res/android"
+                <TableLayout
                     android:id="@+id/confirm_table"
                     android:layout_width="match_parent"
                     android:layout_height="match_parent"
@@ -61,8 +52,21 @@
                 </TableLayout>
 
             </LinearLayout>
-        </LinearLayout>
     </LinearLayout>
+
+            <!--Question-->
+    </LinearLayout>
+
+    <include layout="@layout/multi_question_tab_navigation_buttons"></include>
+    <!--Horizontal Progress bar + text-->
+    <RelativeLayout
+        android:id="@+id/dynamic_footer"
+        android:layout_width="match_parent"
+        android:layout_height="@dimen/footer_navigation_size"
+        android:layout_alignParentBottom="true">
+        <include layout="@layout/progress_bar_row"></include>
+
+    </RelativeLayout>
 
 </RelativeLayout>
