+++ app/src/main/res/layout/multi_question_tab_navigation_buttons.xml
@@ -1,45 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
-<TableRow xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_height="wrap_content"
+<merge xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_height="match_parent"
     android:layout_width="match_parent"
-    android:weightSum="1">
-
-    <LinearLayout
-        android:layout_width="0px"
-        android:layout_height="match_parent"
-        android:orientation="horizontal"
-        android:layout_weight="1"
-        android:weightSum="1"
-        >
-
-        <ImageButton
-            android:id="@+id/left_btn"
-            android:layout_height="wrap_content"
-            android:layout_width="0px"
-            android:layout_weight="0.1"
-            android:layout_gravity="left"
-            android:src="@drawable/back_button"
-            android:background="@color/transparent"
-            android:adjustViewBounds="true"
-            android:scaleType="centerInside"
-            android:enabled="true"
-            />
-        <Space
-            android:layout_width="0dp"
-            android:layout_height="match_parent"
-            android:layout_weight="0.8"/>
-        <ImageButton
-            android:id="@+id/right_btn"
-            android:layout_height="wrap_content"
-            android:layout_width="0px"
-            android:layout_weight="0.1"
-            android:layout_gravity="right"
-            android:src="@drawable/next_button"
-            android:background="@color/transparent"
-            android:adjustViewBounds="true"
-            android:scaleType="centerInside"
-            android:enabled="true"
-            />
-    </LinearLayout>
-
-</TableRow>
+    android:id="@+id/multi_question_tab_navigation_buttons" >
+</merge>
