+++ app/src/main/res/values/dimens.xml
@@ -55,4 +55,8 @@
     <!--Text size for review headers and rows-->
     <dimen name="default_review_fragment_font_size">24dp</dimen>
 
+    <dimen name="footer_navigation_size">30dp</dimen>
+
+
+
 </resources>
