--- app/src/myanmar/java/org/eyeseetea/malariacare/layout/StyleUtils.java
@@ -1,31 +1,0 @@
-package org.eyeseetea.malariacare.layout;
-
-import android.app.Activity;
-import android.content.Context;
-import android.text.Spanned;
-import android.widget.ImageView;
-import android.widget.LinearLayout;
-import android.widget.TabHost;
-import android.widget.TabWidget;
-import android.widget.TextView;
-
-import org.eyeseetea.malariacare.DashboardActivity;
-import org.eyeseetea.malariacare.R;
-import org.eyeseetea.malariacare.database.model.User;
-import org.eyeseetea.malariacare.database.utils.PreferencesState;
-import org.eyeseetea.malariacare.layout.utils.LayoutUtils;
-
-/**
- * Created by idelcano on 29/10/2016.
- */
-
-public class StyleUtils {
-
-    public static void setActionBar(android.support.v7.app.ActionBar actionBar) {
-        LayoutUtils.setActionBarAppAndUser(actionBar);
-    }
-
-    public static void setTabHosts(DashboardActivity dashboardActivity) {
-        dashboardActivity.setTabHostsWithImages();
-    }
-}
