+++ app/src/myanmar/res/layout/assessment_header.xml
@@ -33,6 +33,7 @@
         android:layout_height="fill_parent"
         android:orientation="vertical">
         <View
+            android:layout_marginTop="10dp"
             android:layout_width="match_parent"
             android:layout_height="3dp"
             android:background="@color/dark_violet"/>
