+++ app/src/myanmar/res/layout/dynamic_tab_positiveint_row.xml
@@ -4,6 +4,10 @@
     android:layout_height="match_parent"
     android:layout_width="match_parent"
     android:layout_weight="1">
+    <View
+        android:layout_width="match_parent"
+        android:layout_height="3dp"
+        android:background="@color/light_grey_violet"/>
     <LinearLayout
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
