+++ app/src/myanmar/res/layout/multi_question_tab_navigation_buttons.xml
@@ -1,0 +1,42 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+        android:id="@+id/multi_question_tab_navigation_buttons"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginBottom="@dimen/footer_navigation_size"
+        android:layout_alignParentBottom="true"
+        android:orientation="horizontal"
+        android:weightSum="1"
+        >
+
+        <ImageButton
+            android:id="@+id/left_btn"
+            android:layout_height="wrap_content"
+            android:layout_width="0px"
+            android:layout_weight="0.1"
+            android:layout_gravity="left"
+            android:src="@drawable/back_button"
+            android:background="@color/transparent"
+            android:adjustViewBounds="true"
+            android:scaleType="centerInside"
+            android:enabled="true"
+            />
+
+            <Space
+                android:layout_width="0dp"
+                android:layout_height="match_parent"
+                android:layout_weight="0.8"/>
+
+        <ImageButton
+                android:id="@+id/right_btn"
+                android:layout_height="wrap_content"
+                android:layout_width="0px"
+                android:layout_weight="0.1"
+                android:layout_gravity="right"
+                android:src="@drawable/next_button"
+                android:background="@color/transparent"
+                android:adjustViewBounds="true"
+                android:scaleType="centerInside"
+                android:enabled="true"
+                />
+    </LinearLayout>
