+++ app/src/myanmar/res/layout/question_header.xml
@@ -6,6 +6,7 @@
     android:background="@drawable/background_question_header"
     xmlns:android="http://schemas.android.com/apk/res/android">
     <View
+        android:layout_marginTop="10dp"
         android:layout_width="match_parent"
         android:layout_height="3dp"
         android:background="@color/dark_violet"/>
